requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
axial-positional-embedding==0.2.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
clearml==1.14.4
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
colorlog==6.8.2
commonmark==0.9.1
confection==0.1.4
contractions==0.1.72
cv==1.0.0
cycler==0.12.1
cymem==2.0.8
Cython==0.29.16
docker-pycreds==0.4.0
easydict==1.13
einops==0.7.0
faiss-gpu==1.7.2
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
furl==2.1.3
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
huggingface-hub==0.21.4
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lmdb==1.4.1
local-attention==1.9.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.5.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.2.1
nltk==3.7
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.2.2
onnxruntime==1.17.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.3
pandas==1.4.2
pathlib2==2.3.7.post1
performer-pytorch==1.1.4
pillow==10.2.0
plotly==5.20.0
pluggy==1.4.0
portalocker==2.8.2
preshed==3.0.9
protobuf==3.20.1
py==1.11.0
pyahocorasick==2.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pyDeprecate==0.3.2
PyJWT==2.8.0
pyparsing==3.1.2
pyproject==1.3.1
pyproject-toml==0.0.10
pytest==7.1.2
python-dotenv==0.20.0
pytorch_lightning==1.6.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==12.4.4
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.1
sentry-sdk==1.42.0
setproctitle==1.3.3
sh==1.14.2
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
textsearch==0.0.24
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
transformers==4.39.2
triangle==20230923
triton==2.2.0
typer==0.9.4
urllib3==2.2.1
wandb==0.16.4
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
torch-geometric==2.0.4
